refactor: memberService로 통합
