+++ src/main/java/roomescape/member/service/MemberService.java
@@ -1,12 +1,18 @@
 package roomescape.member.service;
 
 import java.util.List;
+import java.util.Optional;
 import org.springframework.stereotype.Service;
+import roomescape.exception.ConflictException;
 import roomescape.exception.ExceptionCause;
 import roomescape.exception.NotFoundException;
-import roomescape.member.domain.Member;
-import roomescape.member.repository.MemberRepository;
 import roomescape.jwt.TokenProvider;
+import roomescape.member.domain.Member;
+import roomescape.member.dto.LoginRequest;
+import roomescape.member.dto.MemberResponse;
+import roomescape.member.dto.RegistrationRequest;
+import roomescape.member.dto.TokenResponse;
+import roomescape.member.repository.MemberRepository;
 
 @Service
 public class MemberService {
@@ -19,6 +25,22 @@ public class MemberService {
         this.jwtTokenProvider = jwtTokenProvider;
     }
 
+    public TokenResponse createToken(LoginRequest loginRequest) {
+        Optional<Member> memberOptional = memberRepository.findByEmailAndPassword(loginRequest.email(),
+                loginRequest.password());
+        if (memberOptional.isEmpty()) {
+            throw new NotFoundException(ExceptionCause.UNAUTHORIZED_LOGIN_ACCESS);
+        }
+        String token = jwtTokenProvider.createToken(memberOptional.get());
+        return new TokenResponse(token);
+    }
+
+    public MemberResponse findMemberById(Long memberId) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new NotFoundException(ExceptionCause.UNAUTHORIZED_LOGIN_ACCESS));
+        return new MemberResponse(member.getId(), member.getName());
+    }
+
     public List<Member> findAllMembers() {
         return memberRepository.findAll();
     }
@@ -26,5 +48,12 @@ public class MemberService {
     public Member findById(Long memberId) {
         return memberRepository.findById(memberId)
                 .orElseThrow(() -> new NotFoundException(ExceptionCause.UNAUTHORIZED_LOGIN_ACCESS));
+    }
+
+    public void signup(RegistrationRequest registrationRequest) {
+        if (memberRepository.existsByEmail(registrationRequest.email())) {
+            throw new ConflictException(ExceptionCause.MEMBER_DUPLICATE_EMAIL);
+        }
+        memberRepository.save(registrationRequest.createMember());
     }
 }
